Create folderSystem
This class shows a folder system where content is stored as strings. Change the datatype and refactorate the methods to use for other purposes. The object variable perspective is used as reference where options from this “perspective” can be get. If used for the GUI it makes sense as only the requested perspective is loaded. But if performance oriented I would rather use multi dimensional arrays than HashMaps for this purpose. Also are the methods logic used the same. Down side is that amount of  dimensions have to be determined before runtime.

Further explore TreeMap as this class is highly tested. Also you might use a TreeMap as class variable and extend the APIs methods with own object methods. In some case it might be time saving than using this very individualised  class.